pyproject.toml
[tool.poetry]
name = "tiny-dragon"
version = "0.1.0"
description = "are we going to have a mature dragon?"
authors = ["Fatemeh <Fatemeh1725@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
pyyaml = "^6.0.1"
aiohttp-socks = "^0.8.4"
mongoengine = "^0.28.2"
dynaconf = "^3.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
